+++ lib/rubocop/cop/lint/literal_in_interpolation.rb
@@ -25,6 +25,8 @@ module RuboCop
         end
 
         def autocorrect(node)
+          return if node.dstr_type? # nested, fixed in next iteration
+
           value = autocorrected_value(node)
           ->(corrector) { corrector.replace(node.parent.source_range, value) }
         end